alsa: frame_size is size_t, not int
frame_size is a memory size and should be a size_t, not a signed integer.

+++ src/output/alsa_plugin.c
@@ -50,7 +50,7 @@ struct alsa_data {
 	alsa_writei_t *writei;
 	unsigned int buffer_time;
 	unsigned int period_time;
-	int frame_size;
+	size_t frame_size;
 	bool use_mmap;
 
 	struct mixer mixer;
